Code cleanup:
Since we already have the define ISC_STATUS_LENGTH for the the length of
the status vector, we should use it! Replaced 'status[20]' with
'status[ISC_STATUS_LENGTH]'.

Replaced the different symbolic names for the maximum length of a path
(MAX_PATH, MAX_PATH_LENGTH, MAXPATHLEN) with one uniq symbol MAXPATHLEN.


git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@1571 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859