Fix bug: "arithmetic exception, numeric overflow, or string truncation" on VIEW after restore.
This was caused if VIEW1 was referenced in VIEW2, but VIEW2 was restored before VIEW1.
Due that the calculated dbkey_length stored in RDB$RELATIONS for VIEW2 is wrong.
After all relations are restored a recalculation of dbkey_length for VIEWs is done.

This fix should also be added in the first release (SP ?) after FB1.5 final.


git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@10952 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859