Backported a fix to eliminate some wrong length record errors
Normally, deleted records are represented by a stub which has no
data so should not be used for VIO_data calls when building the
version chain.  Under some circumstances, however, the deleted
flag is set on a complete record version with a delta back version.
When that happens, the record should be retrieved


git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/branches/B1_5_Release@24849 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859